add JSON prettying function, add viewuser functionality
Added a function that can take stringifyed JSON objects and put them in a readable format for discord messages.

Added the functionality to view user info with !viewuser, currently it displays a full list of user names with 0 arguments, accepts up to 5 user names to check, returns valid data and specifies invalid data such as user not found or too many users requested.